+++ functions/src/admin-functions/newUser.ts
@@ -1,5 +1,10 @@
 import { onValueCreated } from "firebase-functions/database";
-import { sendEmail } from "../mail/gmail";
+import { GmailConfig, sendEmail } from "../mail/gmail";
+const { defineSecret } = require("firebase-functions/params");
+
+// Define the secret parameter
+const gEmail = defineSecret("GMAIL_EMAIL");
+const gPass = defineSecret("GMAIL_PASSWORD");
 
 
 type Userdata = {
@@ -8,28 +13,40 @@ type Userdata = {
   message: string
 }
 
-const sendNewUserNotificationEmail = onValueCreated("/users/{userId}", (event) => {
-  // …
+const sendNewUserNotificationEmail = onValueCreated({ ref: "/users/{userId}", secrets: [gEmail, gPass] }, (event) => {
 
-
-  const to = "fest@nightofpassion.se";
   const original = event.data.val();
-
   const userdata = { ...original }
   const userId = event.params.userId
 
-  console.log("newUser.ts - event.data ", event.data.val());
-  console.log("newUser.ts - userId ", userId);
-  console.log("newUser.ts - userdata ", userdata);
-  // logger.debug("newUser.ts - event.params.userId: ", event.params.userId);
-  console.log("HELLO new user")
   onNewUser(userId, userdata)
 });
 
 const onNewUser = (userId: string, userdata: Userdata) => {
   console.log("notifyNewUser: ", userId, userdata)
-  // return sendEmail(to, { username: original.name, message: "a new user" })
+
+  const gmailConfig: GmailConfig = {
+    email: gEmail.value(),
+    password: gPass.value()
+  };
+
+  const to = "fest@nightofpassion.se";
+
+  const mailContent = onNewUserEmailTemplate(userdata)
+  return sendEmail(gmailConfig, to, "New NoP signup", mailContent)
 }
 
+type NewUserSignup = {
+  username: string,
+  message: string
+}
+
+const onNewUserEmailTemplate = (data: NewUserSignup): string => {
+  const text = `New NoP signup:
+user: ${data.username || "<unknown user>"}    
+message: ${data.message || "<unknown msg>"}     
+`;
+  return text
+}
 
 export { sendNewUserNotificationEmail };
